CA-51106: ensure that VM.ha_always_run=true means that "VM really should be running. If you see that it isn't running then please start it"
When setting VM.ha_restart_priority, set ha_always_run if the VM is powered on.

This copes with the sequence:
1. start unprotected VM
2. protect VM by setting priority=restart

When shutting down or suspending a VM we unconditionally set ha_always_run to false.

When starting or resuming a VM we set ha_always_run to true if priority=restart and false otherwise.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>