CA-46505: before joining a pool, unplug storage
There's already a constraint there there should be no running VMs... but
there might be disks attached to dom0 and the local SRs are certainly plugged
in. The problem is that pool.join will cause the SR, VDI, PBD, Host XenAPI
references to change and so anyone who might have them cached (e.g. in
a VDI.sm_config lock) will find that they are now invalid.

Therefore we should play safe and unplug the storage before joining the pool.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>