Make bond.create function more efficient
The original code used to call the function Xapi_vm.get_possible_hosts,
which is a very expensive function, more than necessary. This patch
changes that to call the function only when really needed.

This should provide a substantial speed-up in certain cases, especially when
joining a newly installed host to a pool with bonded NICs and many VMs.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>