CA-60270: fix double-close of file descriptor in host internal management network proxy function
The Unixext.proxy convention is: it will close both fds you provide. In this case the 'fromfd' is supposed to be closed within the HTTP handler anyway *but* we must specifically Unix.dup any fd we plan to close ourselves (like the Stunnel one)

Also fix a file descriptor leak if Stunnel.connect fails.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>