[CA-61336] Prevent redo_log from skipping non-consecutive deltas.
The database behaviour has been changed so that all database changes,
not just those which would cause a redo_log write, increase the
generation count. This change means that when reading a database, the
redo_log no longer expects database deltas to be consecutive.

This is a quick fix for beta3 - for RC1 the redo_log should be changed
to expect deltas with a monotonically increasing generation count.

Signed-off-by: John Else <john.else@citrix.com>